Fix eathook bugs (#106)
* Fix infinite loop

* Fix integer undeflow

* Fix integer undeflow

* Fix braces error

+++ sources/EatHook.cpp
@@ -21,7 +21,7 @@ bool PLH::EatHook::hook() {
 	if (pExport == nullptr)
 		return false;
 
-	uint64_t offset = m_fnCallback - m_moduleBase;
+	size_t offset = static_cast<size_t>(m_fnCallback - m_moduleBase);
 
 	/* account for when offset to our function is beyond EAT slots size. We
 	instead allocate a small trampoline within +- 2GB which will do the full
@@ -36,7 +36,7 @@ bool PLH::EatHook::hook() {
 		MemoryProtector protector(m_trampoline, 64, ProtFlag::R | ProtFlag::W | ProtFlag::X, *this, false);
 
 		PLH::ADisassembler::writeEncoding(makeAgnosticJmp(m_trampoline, m_fnCallback), *this);
-		offset = m_trampoline - m_moduleBase;
+		offset = static_cast<size_t>(m_trampoline - m_moduleBase);
 
 		Log::log("EAT hook offset is > 32bit's. Allocation of trampoline necessary", ErrorLevel::INFO);
 	}
